% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnp-package.R
\docType{package}
\name{cnp-package}
\alias{cnp}
\alias{cnp-package}
\title{cnp: Clinical NeuroPsychology normative data methods}
\description{
Functions for psychometric normative data in clinical neuropsychology,
including regression-based adjustment methods (Arcara 2024, Capitani
1987), the Equivalent Scores classification method, and data-simulation
helpers for testing/teaching these methods.
}
\author{
\strong{Maintainer}: Giorgio Arcara \email{giorgio.arcara@gmail.com}

}
\keyword{internal}
