% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TeleGEMS}
\alias{TeleGEMS}
\title{Tele-Global Examination of Mental State (Tele-GEMS) normative sample}
\format{
A data frame with 601 rows and 5 columns:
\describe{
\item{ID}{Sequential participant identifier.}
\item{Age}{Age in years.}
\item{Education}{Years of formal education.}
\item{Sex}{Sex, \code{"M"}/\code{"F"}.}
\item{Tele_GEMS}{Total Tele-GEMS score.}
}
}
\source{
Montemurro, S., Mondini, S., Pucci, V., Durante, G., Riccardi, A.,
Maffezzini, S., Scialpi, G., Signorini, M., & Arcara, G. (2023).
Tele-Global Examination of Mental State (Tele-GEMS): an open tool for the
remote neuropsychological screening. \emph{Neurological Sciences}, 44(10),
3499-3508. \doi{10.1007/s10072-023-06862-1}
}
\usage{
TeleGEMS
}
\description{
Normative data for the Tele-GEMS, a telephone-administered adaptation of
the Global Examination of Mental State (GEMS) developed for remote
neuropsychological screening, collected in an Italian adult sample.
}
\note{
Data licensed under CC BY-NC 4.0 (Attribution-NonCommercial).
}
\keyword{datasets}
